Implemented hashtable generation from symbols (#5)
Hashtable was fixed in code, making it impossible to change the probe
symbol's name used to load the fire function dynamically. A hash table
algorithm was implemented based on how other tools implement it and how
dlfcn loads it. Even though it should work for most cases, there's still
need to implement collision resolution on this algorithm, otherwise,
we'll have problems in the future.

Fixes: https://github.com/sthima/libstapsdt/issues/4