src: add probe arguments
One of the main features of Systap's SDT probes is the ability to pass
arguments to a probe. Those arguments are stored on the processor's
registers, which can be read by other tools through information
available on the SDT note. This commit adds the same functionality,
altough the current implementation only works on x86_64. In the future
we should add support for other processors.

Ref: https://github.com/sthima/libstapsdt/issues/1