src: add probe arguments (#6)
One of the main features of Systemtap's SDT probes is the ability to
pass arguments to a probe. Those arguments are stored in the
processor's registers, which can be read by other tools through
information available on the SDT note. This commit adds the same
functionality, although the current implementation only works on
x86_64. In the future, we should add support for other processors.

Ref: https://github.com/sthima/libstapsdt/issues/1