docs: Add Sphinx to the project
In order to make this library more useful to others, it should provide a
good and updated documentation. Sphinx is a tool to generate docs from
comments in the code, by using ReStructuredText markup language. This
commit adds Sphinx support to the project. Next step is to document the
library API, and then its internals.