src: Remove shared library on provider unload (#18)
Provider unload was correctly cleaning memory and unreferencing symbols,
but the shared library created on load wasn't being removed. We don't
need to keep the shared library, because if the provider is loaded again
it will need to create a new file anyway.

Fixex: https://github.com/sthima/libstapsdt/issues/3