Introduce MemFD_Option_t enum for controlling usage of memfd
This introduces a new enumerated type, MemFD_Option_t, which is used in lieu
of the previous convention of using just an integer type.

All references to these values have been updated, and the symbolic values of
"memfd_enabled" is used in place of the literal "1", and "memfd_disabled" is
used in place of the literel "0".