libstapsdt: support system-wide probing
By adding a USDT probe for stapsdt-created probe firings, we can
enable tools and tracers that understand USDT but do not support
tracing in the dynamically-created shared library, allowing them
to catch dynamic probe firings.  For example, using libbpf, we could
write a BPF program as follows:

SEC("usdt//usr/lib/libstapsdt.so:stapsdt:probe")
int BPF_USDT(args, const char *provider, const char *probename,
	     const char *str, int val)
{
  __bpf_printk("%s/%s fired\n", provider, probename);
  __bpf_printk("got %s, %d args\n", str, val);
  return 0;
}

When attached to the stapsdt:probe in /usr/lib/libstapsdt.so
as above, we can trace events system-wide.

For example, with the above BPF program attached - and running
the example.py from python stapsdt, we see:

 $ cat /sys/kernel/debug/tracing/trace_pipe
         python3-496195  [003] ...11 345760.365976: bpf_trace_printk: pythonapp/firstProbe fired

         python3-496195  [003] ...11 345760.365979: bpf_trace_printk: got My little probe, 42 args

And we will see firings across all processes that utilize
libstapsdt.

Is-enabled support is added via a semaphore for the stapsdt
probe - if the semaphore has a value > 0, it is being traced
and all probes should be considered enabled.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>